simple/train/requirements.txt
Django==1.8.3
django-cors-headers==1.1.0
django-extensions==1.5.5
ipython==3.2.1
psycopg2==2.6.1
pytz==2015.4
requests==2.7.0
six==1.9.0
wheel==0.24.0
xlrd==0.9.4
djangorestframework==3.3.2

